Updated per-process timers to honor the clock ID.
This change really allows differentiation between absolute real time and
absolute monotonic values supplies to timer_settime. The code was previously
treating all absolute times as real time values, which is not correct if the
timer was created for CLOCK_MONOTONIC.

+++ apps/include/libc/sys/types.h
@@ -147,7 +147,7 @@ typedef long ssize_t;
 // Define the type used for timer handles.
 //
 
-typedef int timer_t;
+typedef long timer_t;
 
 //
 // Define a user identifier type.
